implementações diversas on issue #24
